field: fix func_800A5774 STP-bit stride (u_long->u32, same widening class)
Audit follow-up to the FieldImageConvert fix. func_800A5774 (the zoom-fade
semi-transparency pass) ORs 0x80008000 into each VRAM word via a u_long*
work buffer. Its byte count keys off sizeof(u_long) so it doesn't overflow,
but on the 64-bit port each u_long is 8 bytes = 4 pixels while 0x80008000
only covers the low 4 bytes (2 pixels) -- so half the pixels never get their
STP bit set. Type the buffer as u32 (== u_long on PSX, matching-safe) so each
OR covers exactly two 15-bit pixels and the stride stays 4 bytes.

Verified: builds, field still reaches the FieldLoad frontier (no regression).

+++ src/field/main/misc5.c
@@ -66,8 +66,11 @@ void func_800A5774(int x, int y, int h) {
     RECT rect;
     int nSize;
     int i;
-    u_long* pImageBuffer;
-    u_long* pWorkBuffer;
+    /* 15-bit VRAM words: u32 (not u_long) so each |= 0x80008000 covers exactly two
+     * pixels and the stride stays 4 bytes on the 64-bit port. u32 == u_long on PSX,
+     * and nSize keys off sizeof, so this is byte-identical for matching. */
+    u32* pImageBuffer;
+    u32* pWorkBuffer;
 
     rect.x = x;
     rect.y = y;
@@ -76,9 +79,9 @@ void func_800A5774(int x, int y, int h) {
     pImageBuffer = HeapAlloc(h * 0x80, 0x1);
     StoreImage( &rect, pImageBuffer);
     DrawSync(0);
-    
+
     pWorkBuffer = pImageBuffer;
-    nSize = h * (0x80 / sizeof(u_long));
+    nSize = h * (0x80 / sizeof(u32));
     for (i = 0; i < nSize; i += 8) {
         pWorkBuffer[0] |= 0x80008000;
         pWorkBuffer[1] |= 0x80008000;
